apps/api/requirements.txt
fastapi==0.115.11
google-genai>=0.7.0
httpx==0.28.1
Pillow>=11.1.0
pydantic==2.10.6
pydantic-settings>=2.2
pytesseract>=0.3.13
pytest==8.3.5
redis>=5.0
SQLAlchemy>=2.0
psycopg[binary]>=3.2
uvicorn[standard]==0.34.0
